requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.40.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pypdf>=4.2.0
python-docx>=1.1.2
openpyxl>=3.1.5
